+++ package/Editor/Sources/Services/Upm/UpmBaseOperation.cs
@@ -67,7 +67,7 @@ namespace UnityEditor.PackageManager.UI
                 var isVersionCurrent = version == info.version && isCurrent;
                 var isBuiltIn = info.source == PackageSource.BuiltIn;
                 var isVerified = string.IsNullOrEmpty(SemVersion.Parse(version).Prerelease) && version == info.versions.recommended;
-                var state = (isBuiltIn || info.version == lastCompatible) ? PackageState.UpToDate : PackageState.Outdated;
+                var state = (isBuiltIn || info.version == lastCompatible || !isCurrent ) ? PackageState.UpToDate : PackageState.Outdated;
                 
                 // Happens mostly when using a package that hasn't been in production yet.
                 if (info.versions.all.Length <= 0)