+++ package/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [2.0.0] - 2018-06-11
+## [2.0.0] - 2018-07-05
+- Always mark a package as up to date when it's not installed
+- Add "Advanced" menu
+- Hide preview packages by default in stable Unity releases
+- Labelled packages with major version '0' as preview
 - Fixed cropped error message
 - Disabled downgrading Package Manager UI to incompatible versions
 - Updated docs to reflect the new UI